webScraping/requirements.txt
# Web Scraping Dependencies
requests>=2.31.0
urllib3>=2.0.0

# Database (supabase includes postgrest client)
supabase>=2.0.0

# Data Processing
python-dateutil>=2.8.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
responses>=0.14.0
